LayoutTests/imported/w3c/web-platform-tests/encoding/textdecoder-copy.any-expected.txt


FAIL Modify buffer after passing it in assert_equals: expected "@" but got "﻿@"
